fix(langgraph-cli): allow hyphenated prerelease api_version
