+++ .changeset/hunter-allow-js-prerelease.md
@@ -1,0 +1,6 @@
+---
+"@langchain/langgraph-api": patch
+"@langchain/langgraph-cli": patch
+---
+
+fix(langgraph-cli): accept hyphenated prerelease tags in `api_version` values.
