+++ libs/langgraph-api/src/semver/index.mts
@@ -1,7 +1,7 @@
 import * as url from "node:url";
 import * as fs from "node:fs/promises";
 import * as path from "node:path";
-import * as semver from "semver";
+import { satisfiesPeerRange } from "./satisfiesPeerRange.mjs";
 
 const packageJsonPath = path.resolve(
   url.fileURLToPath(import.meta.url),
@@ -23,27 +23,6 @@ export async function checkSemver(
 
     const satisfies = satisfiesPeerRange(pkg.version, required);
     return { ...pkg, required, satisfies };
-  });
-}
-
-function satisfiesPeerRange(version: string, required: string): boolean {
-  if (semver.satisfies(version, required, { includePrerelease: true })) {
-    return true;
-  }
-
-  const parsed = semver.parse(version);
-  if (parsed == null || parsed.prerelease.length === 0) {
-    return false;
-  }
-
-  // CI often installs internal dev builds such as
-  // `1.1.44-dev-<timestamp>`. Semver orders those below the final
-  // `1.1.44`, so `^1.1.44` does not match directly even though the
-  // build targets that release line. Validate the release tuple as a
-  // compatibility check while still rejecting prereleases below range.
-  const releaseVersion = `${parsed.major}.${parsed.minor}.${parsed.patch}`;
-  return semver.satisfies(releaseVersion, required, {
-    includePrerelease: true,
   });
 }
